feat: remote services are no longer managed per owner, but in a centralized manner

+++ server/routers/remote-services.js
@@ -6,7 +6,6 @@ const soasLoader = require('soas')
 const axios = require('axios')
 const requestProxy = require('express-request-proxy')
 const remoteServiceAPIDocs = require('../../contract/remote-service-api-docs')
-const matchstick = require('matchstick')
 const mongoEscape = require('mongo-escape')
 const config = require('config')
 
@@ -18,13 +17,39 @@ const openApiSchema = require('../../contract/openapi-3.0.json')
 openApiSchema.$id = openApiSchema.$id + '-2' // dirty hack to handle ajv error
 const validateOpenApi = ajv.compile(openApiSchema)
 
-const permissions = require('../utils/permissions')
-const usersUtils = require('../utils/users')
 const findUtils = require('../utils/find')
 const asyncWrap = require('../utils/async-wrap')
 const clone = require('fast-clone')
 
-const router = module.exports = express.Router()
+const router = exports.router = express.Router()
+
+// Create default services for the data-fair instance
+exports.init = async(db) => {
+  const remoteServices = db.collection('remote-services')
+  const existingServices = await remoteServices.find().limit(1000).project({ url: 1, id: 1 }).toArray()
+
+  const servicesToAdd = config.remoteServices
+    .filter(s => !existingServices.find(es => es.url === s.url))
+
+  const apisToFetch = new Set(servicesToAdd.map(s => s.url))
+  const apisPromises = [...apisToFetch].map(url => {
+    return axios.get(url)
+      .then(resp => ({ url, api: resp.data }))
+      .catch(err => console.error('Failure to init remote service', err))
+  })
+  const apis = (await Promise.all(apisPromises)).filter(a => a && a.api)
+  const apisDict = Object.assign({}, ...apis.map(a => ({ [a.url]: a.api })))
+  const servicesToInsert = servicesToAdd.filter(s => apisDict[s.url]).map(s => mongoEscape.escape({
+    id: slug(apisDict[s.url].info['x-api-id']),
+    title: apisDict[s.url].info.title,
+    description: apisDict[s.url].info.description,
+    url: s.url,
+    apiDoc: apisDict[s.url],
+    server: apisDict[s.url].servers && apisDict[s.url].servers.length && apisDict[s.url].servers[0].url,
+    actions: computeActions(apisDict[s.url])
+  }, true)).filter(s => !existingServices.find(es => es.id === s.id))
+  if (servicesToInsert.length) await remoteServices.insertMany(servicesToInsert)
+}
 
 // TODO: explain ? simplify ? hard to understand piece of code
 const computeActions = (apiDoc) => {
@@ -42,17 +67,7 @@ const computeActions = (apiDoc) => {
   return actions
 }
 
-const operationsClasses = {
-  list: ['list'],
-  read: ['readDescription', 'readApiDoc'],
-  write: ['writeDescription', 'updateApiDoc'],
-  admin: ['delete', 'getPermissions', 'setPermissions'],
-  use: []
-}
-
 function clean(remoteService) {
-  remoteService.public = permissions.isPublic(remoteService, operationsClasses)
-  delete remoteService.permissions
   delete remoteService._id
   if (remoteService.apiKey && remoteService.apiKey.value) remoteService.apiKey.value = '**********'
   findUtils.setResourceLinks(remoteService, 'remote-service')
@@ -60,6 +75,7 @@ function clean(remoteService) {
 }
 
 // Get the list of remote-services
+// Accessible to anybody
 router.get('', asyncWrap(async(req, res) => {
   const remoteServices = req.app.get('db').collection('remote-services')
   const query = findUtils.query(req, {
@@ -68,7 +84,7 @@ router.get('', asyncWrap(async(req, res) => {
     'api-id': 'apiDoc.info.x-api-id',
     'ids': 'id',
     'id': 'id'
-  })
+  }, true)
   const sort = findUtils.sort(req.query.sort)
   const project = findUtils.project(req.query.select, ['apiDoc'])
   const [skip, size] = findUtils.pagination(req.query)
@@ -78,27 +94,16 @@ router.get('', asyncWrap(async(req, res) => {
   ]
   if (req.query.facets) {
     const q = clone(query)
-    if (req.query.owner) {
-      q.$and.pop()
-      if (!q.$and.length) delete q.$and
-    }
     mongoQueries.push(remoteServices.aggregate(findUtils.facetsQuery(req.query.facets, q)).toArray())
   }
   let [results, count, facets] = await Promise.all(mongoQueries)
-  results.forEach(r => {
-    r.userPermissions = permissions.list(r, operationsClasses, req.user)
-    clean(r)
-  })
+  results.forEach(r => clean(r))
   facets = findUtils.parseFacets(facets)
   res.json({ count, results: results.map(result => mongoEscape.unescape(result, true)), facets })
 }))
 
 const initNew = (req) => {
   const service = { ...req.body }
-  service.owner = usersUtils.owner(req)
-  const date = moment().toISOString()
-  service.createdAt = service.updatedAt = date
-  service.createdBy = service.updatedBy = { id: req.user.id, name: req.user.name }
   if (service.apiDoc) {
     if (service.apiDoc.info) {
       service.title = service.title || service.apiDoc.info.title
@@ -106,16 +111,16 @@ const initNew = (req) => {
     }
     service.actions = computeActions(service.apiDoc)
   }
-  service.permissions = []
   return service
 }
 
-// Create a remote Api
+// Create a remote Api as super admin
 router.post('', asyncWrap(async(req, res) => {
+  if (!req.user) return res.status(401).send()
+  if (!req.user.isAdmin) return res.status(403).send()
   // if title is set, we build id from it
   if (req.body.title && !req.body.id) req.body.id = slug(req.body.title, { lower: true })
   const service = initNew(req)
-  if (!permissions.canDoForOwner(service.owner, 'postRemoteService', req.user, req.app.get('db'))) return res.sendStatus(403)
   if (!validate(service)) return res.status(400).send(ajvErrorMessages(validate.errors))
 
   // Generate ids and try insertion until there is no conflict on id
@@ -137,56 +142,6 @@ router.post('', asyncWrap(async(req, res) => {
   res.status(201).json(clean(service))
 }))
 
-// Create default services for the user and all his organizations, if necessary
-router.post('/_default_services', asyncWrap(async(req, res) => {
-  if (!req.user) return res.sendStatus(401)
-  const remoteServices = req.app.get('db').collection('remote-services')
-  const query = { $or: [{ 'owner.type': 'user', 'owner.id': req.user.id }] }
-  if (req.user.organizations && req.user.organizations.length) {
-    query.$or.push({
-      'owner.type': 'organization',
-      'owner.id': { $in: req.user.organizations.map(o => o.id) }
-    })
-  }
-  const existingServices = await remoteServices.find(query).project({ owner: 1, url: 1, id: 1 }).toArray()
-
-  const owners = [{ type: 'user', id: req.user.id, name: req.user.name }]
-    .concat((req.user.organizations || []).map(o => ({ type: 'organization', id: o.id, name: o.name })))
-  const ownerServices = [].concat(...owners.map(o => config.remoteServices.map(s => ({ owner: o, url: s.href, title: s.title }))))
-
-  const servicesToAdd = ownerServices
-    // exclude service with this owner and url
-    .filter(os => !existingServices.find(es => es.owner.type === os.owner.type && es.owner.id === os.owner.id && es.url === os.url))
-
-  const apisToFetch = new Set(servicesToAdd.map(s => s.url))
-  const apisPromises = [...apisToFetch].map(url => {
-    return axios.get(url)
-      .then(resp => ({ url, api: resp.data }))
-      .catch(err => console.error('Failure to init remote service', err))
-  })
-  const apis = (await Promise.all(apisPromises)).filter(a => a && a.api)
-  const apisDict = Object.assign({}, ...apis.map(a => ({ [a.url]: a.api })))
-  const servicesToInsert = servicesToAdd.filter(s => apisDict[s.url]).map(s => mongoEscape.escape({
-    id: slug(apisDict[s.url].info['x-api-id']) + '-' + s.owner.type + '-' + s.owner.id,
-    title: apisDict[s.url].info.title,
-    description: apisDict[s.url].info.description,
-    url: s.url,
-    apiDoc: apisDict[s.url],
-    server: apisDict[s.url].servers && apisDict[s.url].servers.length && apisDict[s.url].servers[0].url,
-    actions: computeActions(apisDict[s.url]),
-    owner: Object.assign(s.owner.type === 'organization' ? { role: config.adminRole } : {}, s.owner),
-    permissions: (s.owner.type === 'organization' ? [{
-      type: 'organization',
-      id: s.owner.id,
-      name: s.owner.name,
-      classes: ['list', 'read', 'use'],
-      roles: []
-    }] : [])
-  }, true)).filter(s => !existingServices.find(es => es.id === s.id))
-  if (servicesToInsert.length) await remoteServices.insertMany(servicesToInsert)
-  res.status(201).json(`Added ${servicesToInsert.length} services`)
-}))
-
 // Shared middleware
 const readService = asyncWrap(async(req, res, next) => {
   const service = await req.app.get('db').collection('remote-services')
@@ -197,32 +152,23 @@ const readService = asyncWrap(async(req, res, next) => {
   next()
 })
 
-router.use('/:remoteServiceId/permissions', readService, permissions.router('remote-services', 'remoteService'))
-
-// retrieve a remoteService by its id
-router.get('/:remoteServiceId', readService, permissions.middleware('readDescription', 'read'), (req, res, next) => {
-  req.remoteService.userPermissions = permissions.list(req.remoteService, operationsClasses, req.user)
+// retrieve a remoteService by its id as anybody
+router.get('/:remoteServiceId', readService, (req, res, next) => {
   res.status(200).send(clean(req.remoteService))
 })
 
-// PUT used to create or update
+// PUT used to create or update as super admin
 const attemptInsert = asyncWrap(async(req, res, next) => {
+  if (!req.user) return res.status(401).send()
+  if (!req.user.isAdmin) return res.status(403).send()
+
   const newService = initNew(req)
   newService.id = req.params.remoteServiceId
   if (!validate(newService)) return res.status(400).send(ajvErrorMessages(validate.errors))
 
-  // Try insertion if the user is authorized, in case of conflict go on with the update scenario
-  if (permissions.canDoForOwner(newService.owner, 'postRemoteService', req.user, req.app.get('db'))) {
-    try {
-      await req.app.get('db').collection('remote-services').insertOne(mongoEscape.escape(newService, true))
-      return res.status(201).json(clean(newService))
-    } catch (err) {
-      if (err.code !== 11000) throw err
-    }
-  }
   next()
 })
-router.put('/:remoteServiceId', attemptInsert, readService, permissions.middleware('writeDescription', 'write'), asyncWrap(async(req, res) => {
+router.put('/:remoteServiceId', attemptInsert, readService, asyncWrap(async(req, res) => {
   const newService = req.body
   // preserve all readonly properties, the rest is overwritten
   Object.keys(req.remoteService).forEach(key => {
@@ -239,8 +185,11 @@ router.put('/:remoteServiceId', attemptInsert, readService, permissions.middlewa
   res.status(200).json(clean(newService))
 }))
 
-// Update a remote service configuration
-router.patch('/:remoteServiceId', readService, permissions.middleware('writeDescription', 'write'), asyncWrap(async(req, res) => {
+// Update a remote service configuration as super admin
+router.patch('/:remoteServiceId', readService, asyncWrap(async(req, res) => {
+  if (!req.user) return res.status(401).send()
+  if (!req.user.isAdmin) return res.status(403).send()
+
   const patch = req.body
   var valid = validatePatch(patch)
   if (!valid) return res.status(400).send(validatePatch.errors)
@@ -256,8 +205,10 @@ router.patch('/:remoteServiceId', readService, permissions.middleware('writeDesc
   res.status(200).json(clean(mongoEscape.unescape(patchedService)))
 }))
 
-// Delete a remoteService
-router.delete('/:remoteServiceId', readService, permissions.middleware('delete', 'admin'), asyncWrap(async(req, res) => {
+// Delete a remoteService as super admin
+router.delete('/:remoteServiceId', readService, asyncWrap(async(req, res) => {
+  if (!req.user) return res.status(401).send()
+  if (!req.user.isAdmin) return res.status(403).send()
   // TODO : Remove indexes
   await req.app.get('db').collection('remote-services').deleteOne({
     id: req.params.remoteServiceId
@@ -265,7 +216,11 @@ router.delete('/:remoteServiceId', readService, permissions.middleware('delete',
   res.sendStatus(204)
 }))
 
-router.post('/:remoteServiceId/_update', readService, permissions.middleware('updateApiDoc', 'write'), asyncWrap(async(req, res) => {
+// Force update of API definition as super admin
+router.post('/:remoteServiceId/_update', readService, asyncWrap(async(req, res) => {
+  if (!req.user) return res.status(401).send()
+  if (!req.user.isAdmin) return res.status(403).send()
+
   if (!req.remoteService.url) return res.sendStatus(204)
 
   const reponse = await axios.get(req.remoteService.url)
@@ -281,20 +236,8 @@ router.post('/:remoteServiceId/_update', readService, permissions.middleware('up
   res.status(200).json(clean(req.remoteService))
 }))
 
+// Use the proxy as a user with an active session on an application
 router.use('/:remoteServiceId/proxy*', readService, (req, res, next) => {
-  // Match the path with an operation from the doc
-  const operationPath = Object.keys(req.remoteService.apiDoc.paths).filter(path => {
-    const ms = matchstick(path, 'template')
-    return ms.match(req.params['0'])
-  }).map(path => req.remoteService.apiDoc.paths[path])[0]
-
-  // If the operation exists, apply permissions
-  if (operationPath) {
-    const operation = operationPath[req.method.toLowerCase()]
-    if (!permissions.can(req.remoteService, operation.operationId, 'use', req.user)) return res.sendStatus(403)
-  }
-
-  // console.log((req.user && req.user.email) || 'Anonymous', 'is using operation', operation.operationId)
   const options = {
     url: req.remoteService.server + '*',
     headers: { 'x-forwarded-url': `${config.publicUrl}/api/v1/remote-services/${req.remoteService.id}/proxy/` },
@@ -314,15 +257,6 @@ router.use('/:remoteServiceId/proxy*', readService, (req, res, next) => {
   } else if (config.defaultRemoteKey.in === 'header' && config.defaultRemoteKey.value) {
     options.headers[config.defaultRemoteKey.name] = config.defaultRemoteKey.value
   }
-  // transmit organization id as it tends to complement authorization information
-  if (req.remoteService.owner.type === 'organization') {
-    options.headers['x-organizationId'] = req.remoteService.owner.id
-    options.headers['x-organizationName'] = req.remoteService.owner.name
-  }
-  if (req.remoteService.owner.type === 'user') {
-    options.headers['x-userId'] = req.remoteService.owner.id
-    options.headers['x-userName'] = req.remoteService.owner.name
-  }
 
   // We never transmit authentication
   delete req.headers.authorization
@@ -331,6 +265,7 @@ router.use('/:remoteServiceId/proxy*', readService, (req, res, next) => {
   requestProxy(options)(req, res, next)
 })
 
-router.get('/:remoteServiceId/api-docs.json', readService, permissions.middleware('readApiDoc', 'read'), (req, res) => {
+// Anybody can read the API doc
+router.get('/:remoteServiceId/api-docs.json', readService, (req, res) => {
   res.send(req.resourceApiDoc)
 })
